Created a new GUI in order to communicate with the gait selection node properly.
The gait selection now only displays a single gait instead of all of them and the version are loaded every time a gait is selected. The subgait labels change colors when a change is made so the operator has a clear view of the unapplied changes. It is also possible to look at the full version map using a newly created pop up menu.