Created a connection manager in order to establish connection with (multiple) odrives.
The reason why the Json files are stored is that this reduces the connection time of the odrive dramatically since these files do not have to be downloaded when the connection is created. If the found serial number of the odrive is not in the Json files in still downloads the Json file and stores it in the designated directory.