Implement a CI/CD pipeline for the MARCH repository
This commit changes a few things with the goal implementing
a working CI/CD pipeline on GitLab. Previously, the repository used
Travis CI for its CI/CD pipeline, but this has been transfered to
GitLab because we wanted more control over our pipeline.

- Implementation of .gitlab-ci.yml
This commit creates a .gitlab-ci.yml file that divides the
pipeline in 6 stages.

The first stage, "approval" is only available when the .gitlab-ci.yml
has been changed. This stage does nothing, except waiting for approval
from a maintainer.

The second stage, "docker" is only available when the Dockerfile has been
changed. This stage automatically builds a container and uploads it to GitLabs
container registry.

The third stage, "lint" runs some linters that do
not require compilation. Previously, this stage only consisted of
catkin_lint and flake8, but not it also includes autopep8 for broader
coverage of pep8.

The fourth stage, "build" will always run and builds the repository.
This will catch compilation errors in C++, but of course not in Python
files.

The fifth stage, "test" will also always run. It only passes if all
the tests are passing.

The final stage, "static_analysis" will run if a C++ related file has
been changed. This stage runs clang_tidy and depends on the C++
compilation. This is a new stage and may fail (and currently does fail).
Over time, this stage will be required to pass but it is not desirable
to fix this now in this merge request.

- Implementation our own Dockerfile
Previously, the pipeline used ROS Industrial. While this was
relatively easy to setup, it is not perfect for our needs. Our own,
project specific Dockerfile is more desirable here because it creates
a smaller image size.

- Deletion of flaky test
This commit also removes the temperature error tests, because they
turned out to be flaky. This issue has to be addressed and an issue
has been opened on https://gitlab.com/project-march/march/-/issues/375.

- Let linters pass
Since the autopep8 and catkin_lint are more aggressive than before, the
linters that passed before would now fail. These failures have been
fixed in this commit as well.