Disable cert-err58-cpp clang-tidy check
This check results in a lot of errors in code that is not directly
controlled by Project MARCH, such as the gtest suite or the standard
library. Besides that, the dangerous behaviour that cert-err58-cpp tries
to prevent does not seem to brelevant for us, because we do not execute
code before or after main.