Changed and cleaned up State Machine code
TODO/NOTE: Removed std::placeholders::_1/std::placeholders::_2 since these are not used in the state machine implementation. THIS WILL GIVE AN ERROR SOMEWHERE ELSE! This needs to be fixed.

Removed unused imports.
Changed member function and names to be more in line with the conventions we set.
Including adding const refs when necessary.

Also changed the outdated way to pragma once.